refactor(fuse): extract FuseMountArgs into cli module
Move mount CLI parsing and get_conf from the binary into curvine-fuse
cli module without changing runtime mount behavior.
