+++ curvine-fuse/src/cli/mount_args.rs
@@ -1,0 +1,332 @@
+// Copyright 2025 OPPO.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use clap::Parser;
+use curvine_common::conf::ClusterConf;
+use curvine_common::version;
+use orpc::io::net::InetAddr;
+use orpc::{err_box, CommonResult};
+
+/// CLI arguments for the curvine-fuse mount command.
+#[derive(Debug, Parser, Clone)]
+#[command(version = version::VERSION)]
+pub struct FuseMountArgs {
+    // Mount the mount point, mount the file system to a directory of the machine.
+    #[arg(long, help = "Mount point path (default: /curvine-fuse)")]
+    pub mnt_path: Option<String>,
+
+    // Specify the root path of the mount point to access the file system, default "/"
+    #[arg(long, help = "Remote filesystem path (default: /)")]
+    pub(crate) fs_path: Option<String>,
+
+    // Number of mount points
+    #[arg(long, help = "Number of mount points (default: 1)")]
+    pub mnt_number: Option<usize>,
+
+    // Debug mode
+    #[arg(short, long, action = clap::ArgAction::SetTrue, help = "Enable debug mode")]
+    pub(crate) debug: bool,
+
+    // Configuration file path (optional)
+    #[arg(
+        short,
+        long,
+        help = "Configuration file path (optional)",
+        default_value = "conf/curvine-cluster.toml"
+    )]
+    pub(crate) conf: String,
+
+    // IO threads (optional)
+    #[arg(long, help = "IO threads (optional)")]
+    pub io_threads: Option<usize>,
+
+    // Worker threads (optional)
+    #[arg(long, help = "Worker threads (optional)")]
+    pub worker_threads: Option<usize>,
+
+    // How many tasks can read and write data at each mount point
+    #[arg(long, help = "Tasks per mount point (optional)")]
+    pub mnt_per_task: Option<usize>,
+
+    // Whether to enable the clone fd feature
+    #[arg(long, help = "Enable clone fd feature (optional)")]
+    pub clone_fd: Option<bool>,
+
+    // Fuse request queue size
+    #[arg(long, help = "FUSE channel size (optional)")]
+    pub fuse_channel_size: Option<usize>,
+
+    // Read and write file request queue size
+    #[arg(long, help = "Stream channel size (optional)")]
+    pub stream_channel_size: Option<usize>,
+
+    #[arg(long, help = "Enable direct IO (optional)")]
+    pub direct_io: Option<bool>,
+
+    #[arg(long, help = "Cache readdir results (optional)")]
+    pub cache_readdir: Option<bool>,
+
+    // Timeout settings
+    #[arg(long, help = "Entry timeout in seconds (optional)")]
+    pub entry_timeout: Option<f64>,
+
+    #[arg(long, help = "Attribute timeout in seconds (optional)")]
+    pub attr_timeout: Option<f64>,
+
+    #[arg(long, help = "Negative timeout in seconds (optional)")]
+    pub negative_timeout: Option<f64>,
+
+    // Performance settings
+    #[arg(long, help = "Max background operations (optional)")]
+    pub max_background: Option<u16>,
+
+    #[arg(long, help = "Congestion threshold (optional)")]
+    pub congestion_threshold: Option<u16>,
+
+    // Node cache settings
+    #[arg(long, help = "Node cache size (optional)")]
+    pub node_cache_size: Option<u64>,
+
+    #[arg(long, help = "Node cache timeout (e.g., '1h', '30m') (optional)")]
+    pub node_cache_timeout: Option<String>,
+
+    // Fuse web port
+    #[arg(long, help = "Web server port (optional)")]
+    pub web_port: Option<u16>,
+
+    #[arg(long, help = "Master address (e.g., 'm1:8995,m2:8995'")]
+    pub master_addrs: Option<String>,
+
+    // FUSE options
+    #[arg(short, long)]
+    pub(crate) options: Vec<String>,
+
+    // Additional FuseConf fields
+    #[arg(long, help = "Fill inode number when reading directory (optional)")]
+    pub read_dir_fill_ino: Option<bool>,
+
+    #[arg(long, help = "Enable write-back cache (optional)")]
+    pub write_back_cache: Option<bool>,
+
+    #[arg(long, help = "Enable non-seekable mode (optional)")]
+    pub non_seekable: Option<bool>,
+
+    #[arg(long, help = "Enable permission checking (optional)")]
+    pub check_permission: Option<bool>,
+
+    #[arg(long, help = "Enable in-kernel metadata cache (optional)")]
+    pub enable_meta_cache: Option<bool>,
+
+    #[arg(long, help = "Metadata cache capacity in number of entries (optional)")]
+    pub meta_cache_capacity: Option<u64>,
+
+    #[arg(long, help = "Metadata cache TTL (e.g., '120s', '2m') (optional)")]
+    pub meta_cache_ttl: Option<String>,
+
+    #[arg(long, help = "Remember opened inodes across FUSE sessions (optional)")]
+    pub remember: Option<bool>,
+
+    #[arg(long, help = "Auto-cache attr timeout in seconds (optional)")]
+    pub ac_attr_timeout: Option<f64>,
+
+    #[arg(
+        long,
+        help = "Maximum number of entries returned per directory listing (optional)"
+    )]
+    pub list_limit: Option<usize>,
+}
+
+impl FuseMountArgs {
+    /// Parses the cluster configuration file and applies CLI overrides.
+    pub fn get_conf(&self) -> CommonResult<ClusterConf> {
+        let mut conf = match ClusterConf::from(&self.conf) {
+            Ok(c) => {
+                println!("Loaded configuration from {}", self.conf);
+                c
+            }
+            Err(e) => {
+                eprintln!("Warning: Failed to load config file '{}': {}", self.conf, e);
+                eprintln!("Using default configuration");
+                Self::create_default_conf()
+            }
+        };
+
+        // FUSE configuration - only override if command line values are specified
+        if let Some(mnt_path) = &self.mnt_path {
+            conf.fuse.mnt_path = mnt_path.clone();
+        }
+        if let Some(fs_path) = &self.fs_path {
+            conf.fuse.fs_path = fs_path.clone();
+        }
+        if let Some(mnt_number) = self.mnt_number {
+            conf.fuse.mnt_number = mnt_number;
+        }
+        if self.debug {
+            conf.fuse.debug = true;
+        }
+
+        // Optional FUSE parameters - only override if specified
+        if let Some(io_threads) = self.io_threads {
+            conf.fuse.io_threads = io_threads;
+        }
+
+        if let Some(worker_threads) = self.worker_threads {
+            conf.fuse.worker_threads = worker_threads;
+        }
+
+        if let Some(mnt_per_task) = self.mnt_per_task {
+            conf.fuse.mnt_per_task = mnt_per_task;
+        }
+
+        if let Some(clone_fd) = self.clone_fd {
+            conf.fuse.clone_fd = clone_fd;
+        }
+
+        if let Some(fuse_channel_size) = self.fuse_channel_size {
+            conf.fuse.fuse_channel_size = fuse_channel_size;
+        }
+
+        if let Some(stream_channel_size) = self.stream_channel_size {
+            conf.fuse.stream_channel_size = stream_channel_size;
+        }
+
+        if let Some(direct_io) = self.direct_io {
+            conf.fuse.direct_io = direct_io;
+        }
+
+        if let Some(cache_readdir) = self.cache_readdir {
+            conf.fuse.cache_readdir = cache_readdir;
+        }
+
+        if let Some(entry_timeout) = self.entry_timeout {
+            conf.fuse.entry_timeout = entry_timeout;
+        }
+
+        if let Some(attr_timeout) = self.attr_timeout {
+            conf.fuse.attr_timeout = attr_timeout;
+        }
+
+        if let Some(negative_timeout) = self.negative_timeout {
+            conf.fuse.negative_timeout = negative_timeout;
+        }
+
+        if let Some(max_background) = self.max_background {
+            conf.fuse.max_background = max_background;
+        }
+
+        if let Some(congestion_threshold) = self.congestion_threshold {
+            conf.fuse.congestion_threshold = congestion_threshold;
+        }
+
+        if let Some(node_cache_size) = self.node_cache_size {
+            conf.fuse.node_cache_size = node_cache_size;
+        }
+
+        if let Some(node_cache_timeout) = &self.node_cache_timeout {
+            conf.fuse.node_cache_timeout = node_cache_timeout.clone();
+        }
+
+        if let Some(web_port) = self.web_port {
+            conf.fuse.web_port = web_port;
+        }
+
+        if let Some(read_dir_fill_ino) = self.read_dir_fill_ino {
+            conf.fuse.read_dir_fill_ino = read_dir_fill_ino;
+        }
+
+        if let Some(write_back_cache) = self.write_back_cache {
+            conf.fuse.write_back_cache = write_back_cache;
+        }
+
+        if let Some(non_seekable) = self.non_seekable {
+            conf.fuse.non_seekable = non_seekable;
+        }
+
+        if let Some(check_permission) = self.check_permission {
+            conf.fuse.check_permission = check_permission;
+        }
+
+        if let Some(enable_meta_cache) = self.enable_meta_cache {
+            conf.fuse.enable_meta_cache = enable_meta_cache;
+        }
+
+        if let Some(meta_cache_capacity) = self.meta_cache_capacity {
+            conf.fuse.meta_cache_capacity = meta_cache_capacity;
+        }
+
+        if let Some(meta_cache_ttl) = &self.meta_cache_ttl {
+            conf.fuse.meta_cache_ttl = meta_cache_ttl.clone();
+        }
+
+        if let Some(remember) = self.remember {
+            conf.fuse.remember = remember;
+        }
+
+        if let Some(ac_attr_timeout) = self.ac_attr_timeout {
+            conf.fuse.ac_attr_timeout = ac_attr_timeout;
+        }
+
+        if let Some(list_limit) = self.list_limit {
+            conf.fuse.list_limit = list_limit;
+        }
+
+        if let Some(master_addrs) = &self.master_addrs {
+            let mut vec = vec![];
+            for node in master_addrs.split(",") {
+                let tmp: Vec<&str> = node.split(":").collect();
+                if tmp.len() != 2 {
+                    return err_box!("wrong format master_addrs {}", master_addrs);
+                }
+                let hostname = tmp[0].to_string();
+                let port: u16 = tmp[1].parse()?;
+                vec.push(InetAddr::new(hostname, port));
+            }
+            conf.client.master_addrs = vec;
+        }
+
+        // FUSE options - override if provided
+        if !self.options.is_empty() {
+            conf.fuse.fuse_opts = self.options.clone()
+        } else {
+            conf.fuse.fuse_opts = Self::default_mnt_opts();
+        }
+
+        // Re-initialise derived fields (e.g. Duration values) after applying CLI overrides.
+        conf.fuse.init()?;
+
+        Ok(conf)
+    }
+
+    fn create_default_conf() -> ClusterConf {
+        ClusterConf::default()
+    }
+
+    pub fn default_mnt_opts() -> Vec<String> {
+        if cfg!(feature = "fuse3") {
+            vec![
+                "allow_other".to_string(),
+                "async".to_string(),
+                "auto_unmount".to_string(),
+            ]
+        } else {
+            vec![
+                "allow_other".to_string(),
+                "async".to_string(),
+                "direct_io".to_string(),
+                "big_write".to_string(),
+                "max_write=131072".to_string(),
+            ]
+        }
+    }
+}
